// apps/desktop/src/lib/generated/types/EntryId.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type EntryId = "replays" | "addons" | "addons-backups" | "config-backups" | "gameinfo-backups" | "hero-presence-cache" | "shader-cache" | "console-log" | "voice-ban-backups" | "frame-runs" | "settings" | "server-presets" | "replay-rules" | "connection-history" | "notifications" | "patch-notes-index" | "stats-cache" | "server-list-cache" | "replay-info-cache" | "match-cache" | "other-app-files" | "logs";